+++ src/App/Fossa/Config/Report.hs
@@ -19,17 +19,19 @@ import App.Fossa.Config.Common (
  )
 import App.Fossa.Config.ConfigFile (ConfigFile, resolveConfigFile)
 import App.Fossa.Config.EnvironmentVars (EnvVars)
-import App.Fossa.Subcommand (EffStack, GetSeverity (getSeverity), SubCommand (SubCommand))
+import App.Fossa.Subcommand (EffStack, GetCommonOpts (getCommonOpts), GetSeverity (getSeverity), SubCommand (SubCommand))
 import App.Types (BaseDir, OverrideProject (OverrideProject), ProjectRevision)
 import Control.Effect.Diagnostics (Diagnostics, ToDiagnostic (renderDiagnostic), fatal, fromMaybe)
 import Control.Effect.Lift (Has, Lift, sendIO)
 import Control.Timeout (Duration (Seconds))
+import Data.Aeson (ToJSON (toEncoding), defaultOptions, genericToEncoding)
 import Data.List (intercalate)
 import Data.String.Conversion (ToText, toText)
 import Effect.Exec (Exec)
 import Effect.Logger (Logger, Severity (..), pretty)
 import Effect.ReadFS (ReadFS)
 import Fossa.API.Types (ApiOpts)
+import GHC.Generics (Generic)
 import Options.Applicative (
   InfoMod,
   Parser,
@@ -47,7 +49,10 @@ import Options.Applicative (
  )
 import Path.IO (getCurrentDir)
 
-data ReportType = Attribution deriving (Eq, Ord, Enum, Bounded)
+data ReportType = Attribution deriving (Eq, Ord, Enum, Bounded, Generic)
+
+instance ToJSON ReportType where
+  toEncoding = genericToEncoding defaultOptions
 
 instance Show ReportType where
   show Attribution = "attribution"
@@ -56,7 +61,7 @@ data ReportOutputFormat
   = ReportJson
   | ReportMarkdown
   | ReportSpdx
-  deriving (Eq, Ord, Enum, Bounded)
+  deriving (Eq, Ord, Enum, Bounded, Generic)
 
 parseReportOutputFormat :: String -> Maybe ReportOutputFormat
 parseReportOutputFormat s | s == show ReportJson = Just ReportJson
@@ -77,6 +82,9 @@ reportOutputFormatList = intercalate ", " $ map show allFormats
   where
     allFormats :: [ReportOutputFormat]
     allFormats = enumFromTo minBound maxBound
+
+instance ToJSON ReportOutputFormat where
+  toEncoding = genericToEncoding defaultOptions
 
 reportInfo :: InfoMod a
 reportInfo = progDesc desc
@@ -122,6 +130,9 @@ data ReportCliOptions = ReportCliOptions
 
 instance GetSeverity ReportCliOptions where
   getSeverity ReportCliOptions{..} = if (optDebug commons) then SevDebug else SevInfo
+
+instance GetCommonOpts ReportCliOptions where
+  getCommonOpts ReportCliOptions{..} = Just commons
 
 loadConfig ::
   ( Has (Lift IO) sig m
@@ -190,4 +201,7 @@ data ReportConfig = ReportConfig
   , reportType :: ReportType
   , revision :: ProjectRevision
   }
-  deriving (Eq, Ord, Show)
+  deriving (Eq, Ord, Show, Generic)
+
+instance ToJSON ReportConfig where
+  toEncoding = genericToEncoding defaultOptions