+++ src/Strategy/Maven/Pom/Resolver.hs
@@ -6,18 +6,42 @@ module Strategy.Maven.Pom.Resolver (
 ) where
 
 import Algebra.Graph.AdjacencyMap qualified as AM
-import Control.Algebra
-import Control.Carrier.State.Strict
-import Control.Effect.Diagnostics hiding (fromMaybe)
+import Control.Algebra (Has)
+import Control.Carrier.State.Strict (
+  State,
+  get,
+  modify,
+  runState,
+ )
+import Control.Effect.Diagnostics (
+  Diagnostics,
+  context,
+  fatal,
+  recover,
+  (<||>),
+ )
 import Control.Monad (unless)
 import Data.Foldable (traverse_)
 import Data.Map.Strict (Map)
 import Data.Map.Strict qualified as Map
 import Data.Maybe (fromMaybe)
 import Data.Text (Text)
-import Effect.ReadFS
-import Path
-import Strategy.Maven.Pom.PomFile
+import Effect.ReadFS (
+  ReadFS,
+  ReadFSErr (FileReadError),
+  doesFileExist,
+  readContentsXML,
+  resolveDir,
+  resolveFile,
+ )
+import Path (Abs, Dir, File, Path, mkRelFile, parent, (</>))
+import Strategy.Maven.Pom.PomFile (
+  MavenCoordinate,
+  Pom (pomCoord, pomParentCoord),
+  RawParent (rawParentRelativePath),
+  RawPom (rawPomModules, rawPomParent),
+  validatePom,
+ )
 
 data GlobalClosure = GlobalClosure
   { globalGraph :: AM.AdjacencyMap MavenCoordinate
@@ -88,14 +112,14 @@ recursiveLoadPom path = do
 
     recurseRelative :: Text {- relative filepath -} -> m ()
     recurseRelative rel = do
-      resolvedPath :: Maybe (Path Abs File) <- recover $ resolvePath (parent path) rel
+      resolvedPath :: Maybe (Path Abs File) <- recover $ resolvePomPath (parent path) rel
       traverse_ recursiveLoadPom resolvedPath
 
 -- resolve a Filepath (in Text) that may either point to a directory or an exact
 -- pom file. when it's a directory, we default to pointing at the "pom.xml" in
 -- that directory.
-resolvePath :: forall sig m. (Has ReadFS sig m, Has Diagnostics sig m) => Path Abs Dir -> Text -> m (Path Abs File)
-resolvePath cur txt = context "Resolving parent pom.xml path" $ do
+resolvePomPath :: forall sig m. (Has ReadFS sig m, Has Diagnostics sig m) => Path Abs Dir -> Text -> m (Path Abs File)
+resolvePomPath cur txt = context "Resolving parent pom.xml path" $ do
   let resolveToFile :: m (Path Abs File)
       resolveToFile = do
         file <- resolveFile cur txt